Competition validation and cancel fixes (#977)
# Description

## Summary

<!--BRIEF description: DON'T EXPLAIN the code: JUSTIFY what this PR is
for!-->

Allows volunteer validation and auto validate free products
Fix participation cancelling

<!--#### Sources at the end-->

## Issues/PR dependencies

<!--Use a keyword, then #123 for the same repo or
aeecleclair/RepoName#123 for another-->

### Issues to be resolved

<!--Keywords: "closes", "fixes", "resolves" -->
<!--Fixes #-->

### Required PRs

<!--Keywords: "depends on", "blocked by" -->
<!--Depends on #-->

## Changes Made

<!--DESCRIBE the changes: tell the BIG STEPS, use a CHECKLIST to show
progress. You can explain below how the code works.-->

- [x] ...
- [ ] ...

## Additional Notes

<!--Anything relevant that does not quite fit in the summary-->

<!--Don't touch thses two tags-->
<details>
<summary>

# Classification

</summary>

## Type of Change

- [x] 🐛 Bug fix (non-breaking change which fixes an issue)
- [ ] ✨ New feature (non-breaking change which adds functionality)
- [ ] 🔨 Refactor (non-breaking change that neither fixes a bug nor adds
a feature)
- [ ] 🔧 Infra CI/CD (changes to configs of workflows)
- [ ] 💥 BREAKING CHANGE (fix or feature that require a new minimal
version of the front-end)
- [ ] 😶‍🌫️ No impact for the end-users

## Impact & Scope

- [x] Core functionality changes
- [ ] Single module changes
- [ ] Multiple modules changes
- [ ] Database migrations required
- [ ] Other: ... <!--Not module-oriented: write something!-->

## Testing

- [ ] 1. Tested this locally
- [ ] 2. Added/modified tests that pass the CI (or tested in a
downstream fork)
- [ ] 3. Tested in a deployed pre-prod
- [ ] 0. Untestable (exceptionally), will be tested in prod directly

## Documentation

- [ ] Updated [the docs](docs.myecl.fr) accordingly : <!--[Docs#0 -
Title](https://github.com/aeecleclair/myecl-documentation/pull/0)-->
- [ ] `"` Docstrings
- [ ] `#` Inline comments
- [x] No documentation needed

</details>

+++ app/modules/sport_competition/endpoints_sport_competition.py
@@ -788,6 +788,19 @@ async def validate_competition_user(
         edition.id,
         db,
     )
+    purchases = await cruds_sport_competition.load_purchases_by_user_id(
+        user_id,
+        edition.id,
+        db,
+    )
+    for purchase in purchases:
+        if purchase.product_variant.price == 0:
+            await cruds_sport_competition.mark_purchase_as_validated(
+                user.id,
+                purchase.product_variant_id,
+                True,
+                db,
+            )
 
 
 @module.router.patch(
@@ -839,6 +852,7 @@ async def cancel_competition_user(
             participant.team_id,
             db,
         )
+        delete_team = False
         if team is not None:
             if team.captain_id == user_id:
                 next_user = next(
@@ -854,15 +868,17 @@ async def cancel_competition_user(
                         db,
                     )
                 else:
-                    await cruds_sport_competition.delete_team_by_id(
-                        team.id,
-                        db,
-                    )
+                    delete_team = True
         await cruds_sport_competition.delete_participant_by_user_id(
             user_id,
             edition.id,
             db,
         )
+        if delete_team and team is not None:
+            await cruds_sport_competition.delete_team_by_id(
+                team.id,
+                db,
+            )
     await cruds_sport_competition.cancel_competition_user(
         user_id,
         edition.id,
